requirements.txt
pyrogram==2.0.96
tgcrypto
aiohttp
mutagen
psutil
aigpy
python-dotenv
lyricsgenius
pydub
pycryptodomex
Pillow
psycopg2-binary
pathvalidate
music_tag
ffmpeg-python
librespot
python-telegram-bot==20.8
python-dotenv
pyrogram
git+https://github.com/kokarare1212/librespot-python